+++ src/lib.rs
@@ -77,6 +77,7 @@ mod token;
 #[macro_use]
 mod dehtml;
 pub mod html;
+pub mod plaintext;
 
 pub mod dc_receive_imf;
 pub mod dc_tools;
